requirements.txt
Django>=4.0
django-environ
python-dotenv>=0.19.2
